+++ plugins/modules/faas_remove.py
@@ -150,7 +150,7 @@ def main() -> None:
     # check if function is currently deployed and exit early if not
     if function_name := module.params.get('name'):
         if faas.is_deployed(flags=flags, args=flags_args[1], name=function_name) is False:
-            module.log(msg=f'Function {function_name} is not deployed, skipping removal')
+            module.debug(msg=f'Function {function_name} is not deployed, skipping removal')
             module.exit_json(changed=False, command=command)
 
     # execute faas